Correctly use withPropertyName() in Gradle 4.3
This method was moved from TaskPropertyBuilder to TaskFilePropertyBuilder
in Gradle 4.3.

Fixes #66
